% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{addIndexMeta}
\alias{addIndexMeta}
\alias{addIndexMeta.classic}
\alias{addIndexMeta,ANY-method}
\title{Add necessary elements of the metadata slot of an altervnative splicing SCFind index object}
\usage{
addIndexMeta.classic(object, type, read_count, stats, node_list)

\S4method{addIndexMeta}{ANY}(object, type, read_count, stats, node_list)
}
\arguments{
\item{object}{an SCFind class object built by the method "builtAltSpliceIndex"}

\item{type}{a character to indicate the type and node contents of the index}

\item{read_count}{the read count matrix built by the function "buildMatrix.read_count"}

\item{stats}{the statistics matrix built by the function "buildMatrix.stats"}

\item{node_list}{the node information matrix built by the function "buildMatrix.node_list"}
}
\value{
an SCFind object with classic metadata components
}
\description{
Add necessary elements of the metadata slot of an altervnative splicing SCFind index object
}
