% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{buildAltSpliceIndex}
\alias{buildAltSpliceIndex}
\alias{buildAltSpliceIndex.PSI}
\alias{buildAltSpliceIndex,ANY-method}
\title{Builds an \code{SCFind} object from a \code{matrix}}
\usage{
buildAltSpliceIndex.PSI(psival, metadata, dataset.name, column.label, qb = 2)

\S4method{buildAltSpliceIndex}{ANY}(
  psival,
  metadata,
  dataset.name,
  column.label = "primary_type",
  qb = 2
)
}
\arguments{
\item{psival}{a dataframe object containing psi value of each node of a gene and cell with gene_node as row.names and cell id as col.names}

\item{metadata}{metadata of the alternative splicing dataset, must include cell.type information for the dataset}

\item{dataset.name}{name of the dataset, i.e. tissue name}

\item{column.label}{the cell.type in metadata that will be used for the index}

\item{qb}{number of bits per cell that are going to be used for quantile compression of the expression data}
}
\value{
an SCFind object
}
\description{
This function will index a \code{matrix} including alternative splicing information for each cell as an SCFind index.
}
