% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildMatrix_new.R
\name{buildMatrix.original}
\alias{buildMatrix.original}
\title{Builds an original \code{matrix} from a given Whippet Collapsed Quant .psi.gz files containing directory and a metadata}
\usage{
buildMatrix.original(file)
}
\arguments{
\item{file}{the merged Whippet .psi.tsv file, can be an object or a url}
}
\value{
a tibble object for further scailing
}
\description{
Builds an original \code{matrix} from a given Whippet Collapsed Quant .psi.gz files containing directory and a metadata
}
