% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildMatrix.R
\name{buildMatrix.read_count}
\alias{buildMatrix.read_count}
\title{Builds a read_count \code{matrix} from a given Whippet Quant .psi.gz files containing directory and a metadata 
The read_count matrix would be stored in the alternative splicing index metadata slot}
\usage{
buildMatrix.read_count(dir, metadata)
}
\arguments{
\item{dir}{a directory containing .psi.gz files from Whippet quantification}

\item{metadata}{metadata of the alternative splicing dataset, must include a column named SRA and has SRA ids for all Whippet quantified samples}
}
\value{
a matrix object(read count)
}
\description{
Builds a read_count \code{matrix} from a given Whippet Quant .psi.gz files containing directory and a metadata 
The read_count matrix would be stored in the alternative splicing index metadata slot
}
