% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildMatrix.R
\name{buildMatrix.scaled.above}
\alias{buildMatrix.scaled.above}
\title{Scale the original matrix and build the input matrix of an above index
The output of this function correspond to the input param "psival" in the buildAltSpliceIndex method}
\usage{
buildMatrix.scaled.above(matrix.original)
}
\arguments{
\item{matrix.original}{the output original matrix from the function "buildMatrix.original"}
}
\value{
a matrix object(scaled above)
}
\description{
Scale the original matrix and build the input matrix of an above index
The output of this function correspond to the input param "psival" in the buildAltSpliceIndex method
}
