% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildMatrix_new.R
\name{buildMatrix.stats}
\alias{buildMatrix.stats}
\title{Calculate dataset-wide mean and SD of psi values of all nodes
The stats matrix would be stored in the alternative splicing index metadata slot}
\usage{
buildMatrix.stats(matrix.original, matrix.scaled)
}
\arguments{
\item{matrix.original}{the output original matrix from the function "buildMatrix.original"}

\item{matrix.scaled}{the output scaled matrix from either the function "buildMatrix.scaled.above" or "buildMatrix.scaled.below"}
}
\value{
a matrix object(stats)
}
\description{
Calculate dataset-wide mean and SD of psi values of all nodes
The stats matrix would be stored in the alternative splicing index metadata slot
}
