% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{findMutuallyExclusive}
\alias{findMutuallyExclusive}
\alias{find.mutually.exclusive}
\alias{findMutuallyExclusive,SCFind-method}
\title{This function finds coordinated node sets for a gene}
\usage{
find.mutually.exclusive(object)

\S4method{findMutuallyExclusive}{SCFind}(object)
}
\arguments{
\item{object}{the \code{SCFind} object}
}
\value{
a dataframe that contains potential mutually exclusive nodes in the index
}
\description{
This function finds coordinated node sets for a gene
}
