% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{findNodeSets}
\alias{findNodeSets}
\alias{gene.node.sets}
\alias{findNodeSets,SCFind,character,character-method}
\title{This function finds coordinated node sets for a gene}
\usage{
gene.node.sets(object, gene.list, query.type)

\S4method{findNodeSets}{SCFind,character,character}(object, gene.list, query.type)
}
\arguments{
\item{object}{the \code{SCFind} object}

\item{gene.list}{gene id or gene name to find coordinated node sets}

\item{query.type}{either "gene_id" or "external_gene_name" to use in query}
}
\value{
a dataframe that contains nodes for gene.list
}
\description{
This function finds coordinated node sets for a gene
}
