% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{findSimilarGenes}
\alias{findSimilarGenes}
\alias{similar.nodes}
\title{Look at all other genes and rank them based on the similarity of their expression pattern to the pattern defined by the gene query}
\usage{
similar.nodes(object, gene.list, datasets, top.k = 5)
}
\arguments{
\item{object}{the \code{SCFind} object}

\item{gene.list}{genes to be searched in the gene.index}

\item{datasets}{the datasets that will be considered}

\item{top.k}{how many genes to retrieve}
}
\value{
the list of genes and their similarities presented in Jaccard indices
}
\description{
Look at all other genes and rank them based on the similarity of their expression pattern to the pattern defined by the gene query
}
