% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/CoreMethods.R
\name{findSimilarNodes}
\alias{findSimilarNodes}
\alias{findSimilarNodes,SCFind,character-method}
\title{Look at all other genes and rank them based on the similarity of their expression pattern to the pattern defined by the gene query}
\usage{
findSimilarNodes(object, gene.list, datasets, top.k = 5)

\S4method{findSimilarNodes}{SCFind,character}(object, gene.list, datasets, top.k = 5)
}
\description{
Look at all other genes and rank them based on the similarity of their expression pattern to the pattern defined by the gene query
}
