% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{geneNodes}
\alias{geneNodes}
\alias{gene.nodes}
\alias{geneNodes,SCFind,character,character-method}
\title{This function retrieves node details in an index by gene id or gene name}
\usage{
gene.nodes(object, gene.list, query.type)

\S4method{geneNodes}{SCFind,character,character}(object, gene.list, query.type)
}
\arguments{
\item{object}{the \code{SCFind} object}

\item{gene.list}{gene id or gene name list to find nodes}

\item{query.type}{either "Gene", "external_gene_name", "Gene_node" to use in query}
}
\value{
a dataframe that contains nodes for gene.list
}
\description{
This function retrieves node details in an index by gene id or gene name
}
