% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreFunctions.R
\name{get.cell.type.raw.psi}
\alias{get.cell.type.raw.psi}
\title{This function gets the raw PSI of a node in a cell type}
\usage{
get.cell.type.raw.psi(object, gene.list, cell.type, index.type)
}
\arguments{
\item{object}{the \code{SCFind} object}

\item{gene.list}{several nodes that we wish to get the raw PSI}

\item{cell.type}{cell type tp query, can only query one cell type at once}

\item{index.type}{above or below to indicate the type of splicing index}
}
\value{
a dataframe that contains raw psi value in the queried cell type of the gene.list
}
\description{
This function gets the raw PSI of a node in a cell type
}
