% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreFunctions.R
\name{get_lower_tri}
\alias{get_lower_tri}
\title{Get lower triangle of the correlation matrix}
\usage{
get_lower_tri(cormat)
}
\arguments{
\item{cormat}{correlation matrix}
}
\description{
Get lower triangle of the correlation matrix
}
