% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreFunctions.R
\name{get_upper_tri}
\alias{get_upper_tri}
\title{# Get upper triangle of the correlation matrix}
\usage{
get_upper_tri(cormat)
}
\arguments{
\item{cormat}{correlation matrix}
}
\description{
# Get upper triangle of the correlation matrix
}
