% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{getCoordinatedNodes}
\alias{getCoordinatedNodes}
\alias{get_coordinated_nodes}
\alias{getCoordinatedNodes,SCFind,character-method}
\title{This function finds coordinated node sets for a gene}
\usage{
get_coordinated_nodes(object, gene.name)

\S4method{getCoordinatedNodes}{SCFind,character}(object, gene.name)
}
\arguments{
\item{object}{the \code{SCFind} object}

\item{gene.name}{Gene_name of the interested gene from geneNodes}
}
\value{
a dataframe that contains top queries
}
\description{
This function finds coordinated node sets for a gene
}
