% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{hyperQueryCellTypesAS}
\alias{hyperQueryCellTypesAS}
\alias{cell.types.phyper.test.AS}
\alias{hyperQueryCellTypesAS,SCFind,character-method}
\title{Runs a query and performs the hypergeometric test for the retrieved cell types}
\usage{
cell.types.phyper.test.AS(object, node.list, datasets)

\S4method{hyperQueryCellTypesAS}{SCFind,character}(object, node.list, datasets)
}
\arguments{
\item{object}{the \code{SCFind} object}

\item{node.list}{AS nodes to be searched in the node.list.index
(Operators: "-gene" to exclude a gene | "*gene" either gene is expressed
"*-gene" either gene is expressed to be excluded)}

\item{datasets}{the datasets vector that will be tested as background for the hypergeometric test}
}
\value{
a DataFrame that contains all cell types with the respective cell cardinality and the hypergeometric test
}
\description{
Runs a query and performs the hypergeometric test for the retrieved cell types
}
