% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{nodeDetails}
\alias{nodeDetails}
\alias{node.details}
\alias{nodeDetails,SCFind,character-method}
\title{This function retrieves node details by node id}
\usage{
node.details(object, node.list)

\S4method{nodeDetails}{SCFind,character}(object, node.list)
}
\arguments{
\item{object}{the \code{SCFind} object}

\item{node.list}{AS node ids to find their details}
}
\value{
a dataframe that contains details of the query nodes
}
\description{
This function retrieves node details by node id
}
