% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{plotRawPsiCorr}
\alias{plotRawPsiCorr}
\alias{plot.raw.psi.corr}
\alias{plotRawPsiCorr,data.frame,character,character-method}
\title{This function plots correlation of splicing PSI and return significantly correlated nodes}
\usage{
\method{plot}{raw.psi.corr}(raw_psi, node.list = "all_nodes", cell.types = "all_cell_types")

\S4method{plotRawPsiCorr}{data.frame,character,character}(raw_psi, node.list = "all_nodes", cell.types = "all_cell_types")
}
\arguments{
\item{raw_psi}{raw psi matrx, each row is a gene, each col is a cell}

\item{node.list}{node list whose PSI is to be plotted, default 'all_nodes'}

\item{cell.types}{cell types to consider, default 'all_cell_types'}
}
\value{
a list object with heatmap, pos corr and neg corr
}
\description{
This function plots correlation of splicing PSI and return significantly correlated nodes
}
