% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildMatrix_new.R
\name{scaleMatrix.diff}
\alias{scaleMatrix.diff}
\title{Scale the original \code{matrix} with difference from dataset mean
This version is stringent that it requires the difference of PSI to mean is higher than 0.2}
\usage{
scaleMatrix.diff(matrix.original)
}
\arguments{
\item{matrix.original}{the original matrix from buildMatrix.original}
}
\value{
a scaled matrix object
}
\description{
Scale the original \code{matrix} with difference from dataset mean
This version is stringent that it requires the difference of PSI to mean is higher than 0.2
}
