% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildMatrix_new.R
\name{scaleMatrix.z}
\alias{scaleMatrix.z}
\title{Scale the original \code{matrix} with Z-score normalization}
\usage{
scaleMatrix.z(matrix.original)
}
\arguments{
\item{matrix.original}{the original matrix from buildMatrix.original}
}
\value{
a scaled matrix object
}
\description{
Scale the original \code{matrix} with Z-score normalization
}
