% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{scfindNodes}
\alias{scfindNodes}
\alias{scfind.get.nodes.in.db}
\alias{scfindNodes,SCFind-method}
\title{Get all nodes in the database}
\usage{
scfind.get.nodes.in.db(object)

\S4method{scfindNodes}{SCFind}(object)
}
\arguments{
\item{object}{the \code{scfind} object}
}
\value{
the list of genes present in the database
}
\description{
Get all nodes in the database
}
