% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tasic_2016.R
\docType{data}
\name{Tasic_2016}
\alias{Tasic_2016}
\title{Example SCFind alternative splicing index
`Tasic_2016` Example SCFind alternative splicing index data for mouse glutamatergic and GABA-ergic neurons}
\format{
url to retrieve SCfind indexes.
}
\source{
\url{/lustre/scratch117/cellgen/team218/ys6/Tasic_data}
}
\usage{
data(Tasic_2016)
}
\description{
Example SCFind alternative splicing index
`Tasic_2016` Example SCFind alternative splicing index data for mouse glutamatergic and GABA-ergic neurons
}
\references{
Allan Brain Atlas (https://www.nature.com/articles/nn.4216)).
}
\keyword{index}
