% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{addIndexMeta}
\alias{addIndexMeta}
\alias{addIndexMeta.classic}
\alias{addIndexMeta,ANY-method}
\title{Add necessary elements of the metadata slot of an altervnative splicing SCFind index object}
\usage{
addIndexMeta.classic(object, stats, node_list, diff_cut)

\S4method{addIndexMeta}{ANY}(object, stats, node_list, diff_cut)
}
\arguments{
\item{object}{an SCFind class object built by the method "builtAltSpliceIndex"}

\item{stats}{the statistics matrix built by the function "buildMatrix.stats"}

\item{node_list}{the node information matrix built by the function "buildMatrix.node_list"}

\item{diff_cut}{sparse matrix with 1 denoting the removal of node PSI due to not sufficiently deviate from tissue mean}
}
\value{
an SCFind object with classic metadata components
}
\description{
Add necessary elements of the metadata slot of an altervnative splicing SCFind index object
}
