% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildMatrix_new.R
\name{buildMatrix.above}
\alias{buildMatrix.above}
\title{Select the desired type of splicing events nodes from the scaled matrix
Above dataset mean inclusion events}
\usage{
buildMatrix.above(matrix.scaled)
}
\arguments{
\item{matrix.scaled}{the output of either scaleMatrix.diff or scaleMatrix.z}
}
\value{
a matrix object
}
\description{
Select the desired type of splicing events nodes from the scaled matrix
Above dataset mean inclusion events
}
