% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildMatrix_new.R
\name{buildMatrix.node_list}
\alias{buildMatrix.node_list}
\title{Collect details of all nodes, including gene id, gene name, node type, strand, coordinates etc.
The node_list matrix would be stored in the alternative splicing index metadata slot}
\usage{
buildMatrix.node_list(dir, matrix.scaled)
}
\arguments{
\item{dir}{a directory containing .psi.gz files from Whippet quantification}

\item{matrix.scaled}{the output scaled matrix from either the function "buildMatrix.scaled.above" or "buildMatrix.scaled.below"}
}
\value{
a matrix object(node_list)
}
\description{
Collect details of all nodes, including gene id, gene name, node type, strand, coordinates etc.
The node_list matrix would be stored in the alternative splicing index metadata slot
}
