% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildMatrix.R
\name{buildMatrix.original}
\alias{buildMatrix.original}
\title{Builds an original \code{matrix} from a given Whippet Quant .psi.gz files containing directory and a metadata}
\usage{
buildMatrix.original(dir, metadata)
}
\arguments{
\item{dir}{a directory containing .psi.gz files from Whippet quantification, looks like "/lustre/scratch117/cellgen/team218/ys6/TM_FACS/MicroExonator/Whippet/Quant/Single_Cell/Unpooled/"}

\item{metadata}{metadata of the alternative splicing dataset, must include a column named "SRA" and has SRA ids for all Whippet quantified samples}
}
\value{
a matrix object(original)
}
\description{
Builds an original \code{matrix} from a given Whippet Quant .psi.gz files containing directory and a metadata
}
