% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/CoreMethods.R
\name{findHouseKeepingNodes}
\alias{findHouseKeepingNodes}
\alias{findHouseKeepingNodes,SCFind,character-method}
\alias{findHouseKeepinNodes}
\title{Find the set of genes that are ubiquitously expressed in a query of cell types}
\usage{
findHouseKeepingNodes(object, cell.types, min.recall = 0.5, max.genes = 1000)

\S4method{findHouseKeepingNodes}{SCFind,character}(object, cell.types, min.recall = 0.5, max.genes = 1000)
}
\description{
Find the set of genes that are ubiquitously expressed in a query of cell types
}
