% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{findMutuallyExclusive}
\alias{findMutuallyExclusive}
\alias{find.mutually.exclusive}
\alias{findMutuallyExclusive,SCFind,character-method}
\title{This function finds coordinated node sets for a gene}
\usage{
find.mutually.exclusive(object, node.types)

\S4method{findMutuallyExclusive}{SCFind,character}(object, node.types)
}
\arguments{
\item{object}{the \code{SCFind} object}

\item{node.types}{the types of nodes to find mutually exclisive events in}
}
\value{
a dataframe that contains potential mutually exclusive nodes in the index
}
\description{
This function finds coordinated node sets for a gene
}
