% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/CoreMethods.R
\name{findNodeSignatures}
\alias{findNodeSignatures}
\alias{findNodeSignatures,SCFind-method}
\title{Find the signature genes for a cell-type}
\usage{
findNodeSignatures(
  object,
  cell.types,
  max.genes = 1000,
  min.cells = 10,
  max.pval = 0
)

\S4method{findNodeSignatures}{SCFind}(
  object,
  cell.types,
  max.genes = 1000,
  min.cells = 10,
  max.pval = 0
)
}
\description{
Find the signature genes for a cell-type
}
