% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{getRawPsi}
\alias{getRawPsi}
\alias{get.raw.psi}
\alias{getRawPsi,SCFind,character,character-method}
\title{This function gets the raw PSI of a node in a cell type}
\usage{
get.raw.psi(object, node.list, cell.types)

\S4method{getRawPsi}{SCFind,character,character}(object, node.list, cell.types)
}
\arguments{
\item{object}{the \code{SCFind} object}

\item{node.list}{several nodes that we wish to get the raw PSI}

\item{cell.types}{cell type tp query, can only query one cell type at once}
}
\value{
a dataframe that contains raw psi value in the queried cell type of the gene.list
}
\description{
This function gets the raw PSI of a node in a cell type
}
