% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{plotRawPsiHeatmap}
\alias{plotRawPsiHeatmap}
\alias{plot.raw.psi.heatmap}
\alias{plotRawPsiHeatmap,SCFind,character,character,character-method}
\title{This function plots heatmap of PSI values}
\usage{
\method{plot}{raw.psi.heatmap}(object, node.list, cell.types, index.type)

\S4method{plotRawPsiHeatmap}{SCFind,character,character,character}(object, node.list, cell.types, index.type)
}
\arguments{
\item{object}{the \code{SCFind} object}

\item{node.list}{a list of nodes whose raw psi is to be plotted}

\item{cell.types}{a list of cell types whose raw psi is to be plotted}

\item{index.type}{above or below as the type of the input index}
}
\description{
This function plots heatmap of PSI values
}
